Price/Balance job, Market Repo (#6)
* balance,price jobs added

* pg market db impl

* markets loader, price/balance fetcher impl

* make github api url configurable

* add docker compose, make dev

* error handling, logging interceptors

* test fixes

* change registry url

* docker setup update

* fetchPrice/fetchBalance update

* handle panic

* fix test

+++ internal/core/domain/market_price.go
@@ -4,9 +4,9 @@ import "time"
 
 type MarketPrice struct {
 	MarketID   string
-	BasePrice  int
+	BasePrice  float32
 	BaseAsset  string
-	QuotePrice int
+	QuotePrice float32
 	QuoteAsset string
 	Time       time.Time
 }
